refactor(trip-details): continue organizing components into folder
