+++ packages/trip-details/src/types.ts
@@ -52,6 +52,22 @@ export interface FareDetailsProps {
   minTNCFare: number;
 }
 
+export interface FareConfig {
+  /**
+   * Determines which transit fare should be displayed by default, should there be multiple transit fare types.
+   */
+  defaultFareType?: { headerKey: string } & FareProductSelector;
+  /**
+   * Column and table configuration for fare details/fare by leg table.
+   */
+  fareDetailsLayout?: FareTableLayout[];
+  /**
+   * Mapping between fare keys and human-readable names for them.
+   */
+  fareKeyNameMap?: {
+    [name: string]: string;
+  };
+}
 
 export interface TripDetailsProps {
   /**
@@ -63,32 +79,17 @@ export interface TripDetailsProps {
    */
   co2Config?: CO2ConfigType;
   /**
-   * Determines which transit fare should be displayed by default, should there be multiple transit fare types.
-   */
-  defaultFareType?: { headerKey: string } & FareProductSelector;
-  /**
    * Slot for a custom component to render the expandable section for departure.
    */
   DepartureDetails?: React.ElementType<DepartureDetailsProps>;
-  /**
-   * If this is set to true, a row will be added to the trip details displaying how
-   * many minutes in total the user will spend walking or biking.
-   */
-  displayTimeActive?: boolean;
   /**
    * Slot for a custom component to render the expandable section for fares.
    */
   FareDetails?: React.ElementType<FareDetailsProps>;
   /**
-   * Column and table configuration for fare details/fare by leg table.
+   * Configuration for fare details/fare by leg table.
    */
-  fareDetailsLayout?: FareTableLayout[];
-  /**
-   * Mapping between fare keys and human-readable names for them.
-   */
-  fareKeyNameMap?: {
-    [name: string]: string;
-  };
+  fareConfig?: FareConfig;
   /**
    * Itinerary that the user has selected to view, contains multiple legs.
    */
